pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "spider"
version = "1.0.0"
description = "System for Penetration Testing, Intelligence, Discovery, Exploit & Recon"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Shubham" }]
keywords = ["cybersecurity", "pentesting", "llm", "langgraph", "metasploit", "offensive-security"]
dependencies = [
    "langgraph==0.2.28",
    "langchain-core==0.3.15",
    "openai==1.51.0",
    "sqlalchemy==2.0.36",
    "pymetasploit3==1.0.3",
    "rich==13.9.4",
    "click==8.1.8",
    "jinja2==3.1.4",
    "weasyprint==62.3",
    "requests==2.32.3",
    "python-dotenv==1.0.1",
    "pydantic==2.9.2",
    "python-libnmap==0.7.3",
]

[project.scripts]
spider = "spider.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["spider*"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"
